contact widget: Inherit from GtkListBoxRow
This is more practical than inheriting from GtkGrid. I first thought we
might need this widget in other places than the list boxes, but it seems
this won't be the case. (Except for the me button, but the fact that
the widget is now a list box row doesn't break it there.)