contact row: Add API for the "selected" property
Add a toggle function for the property. Also select/unselect rows when
activating them (clicking anywhere, not just the check button) when in
selection mode and unselect all rows when leaving the selection mode.

Note: The GTK_SELECTION_MULTIPLE selection mode for GtkListBox is
currently not usable here, as it 1) is currently broken for pointer
devices and 2) doesn't offer check boxes by default anyway. So we keep
the GTK_SELECTION_SINGLE selection mode and implement multiple selection
with the use of the "selected" property.