configure.ac: Remove two AC_PATH_PROG checks
glib-compile-resources and glib-genmarshal should come with the
glib2-devel package, which we are already checking for. Besides, using
AC_PATH_PROG without testing for its result will only cause a cryptic
error during make in case the programs are missing. Now we get a clear
command not found message instead.