file transfer: transferred-size property changes
Remove the _set_ function and replace it by an _add_ function, and put
the calculations whether to notify about the property change there. We
currently notify on every 1000th part of the file that has been
transmitted. Now we consistently print debug messages for the progress
of incoming and outgoing file transfers.