window: The comeback of Accept and Discard buttons
Show Accept and Discard buttons in the header bar when we show requests
and the selection mode is not enabled. Switching to selection mode to
accept a single requests seemed a bit too cumbersome. The buttons use
two new window actions: "accept-active" and "delete-active".

This commit also ensures that we don't show any other button than the
selection button in the header bar when we are in selection mode.